+++ src/web-configuration.ts
@@ -1,12 +1,13 @@
-import { window } from 'vscode';
+import { commands, window } from 'vscode';
 import { Project } from './project';
 import { getSetting, setSetting, WorkspaceSetting } from './workspace-state';
+import { Context, VSCommand } from './context-variables';
 
 export enum WebConfigSetting {
-  welcome = 'Show preview and open web browser',
-  welcomeNoBrowser = 'Show preview without opening browser',
-  browser = 'Open web browser',
-  editor = 'Open app in editor',
+  nexus = 'WebConfigNexusBrowser',
+  browser = 'WebConfigWebBrowser',
+  editor = 'WebConfigEditor',
+  none = 'WebConfigNone',
 }
 
 export function getWebConfiguration(): WebConfigSetting {
@@ -14,31 +15,11 @@ export function getWebConfiguration(): WebConfigSetting {
   if (setting) {
     return setting;
   } else {
-    return WebConfigSetting.welcome;
+    return WebConfigSetting.browser;
   }
 }
 
-export async function webConfiguration(project: Project): Promise<string | undefined> {
-  const setting = getSetting(WorkspaceSetting.webAction);
-  const configs = [
-    check(WebConfigSetting.welcome, setting),
-    check(WebConfigSetting.welcomeNoBrowser, setting),
-    check(WebConfigSetting.browser, setting),
-    check(WebConfigSetting.editor, setting),
-  ];
-
-  const selection = await window.showQuickPick(configs, {
-    placeHolder: 'Select the default action when running for web',
-  });
-  if (selection) {
-    setSetting(WorkspaceSetting.webAction, selection);
-  }
-  return selection;
-}
-
-function check(msg: string, setting: string): string {
-  if (msg === setting) {
-    return msg + ` $(check)`;
-  }
-  return msg;
+export async function setWebConfig(setting: WebConfigSetting) {
+  setSetting(WorkspaceSetting.webAction, setting);
+  commands.executeCommand(VSCommand.setContext, Context.webConfig, setting);
 }